firebase_datastore/requirements.txt
google-cloud-datastore=1.6.0
